app/requirements.txt
altgraph==0.17.4
backports.tarfile==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
comtypes==1.4.12
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
fonttools==4.60.1
idna==3.10
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
kiwisolver==1.4.9
matplotlib==3.10.6
more-itertools==10.8.0
numpy==2.3.3
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pipreqs==0.4.13
psutil==7.1.0
pycaw==20240210
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
pystray==0.19.5
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
u==1.0.4
urllib3==2.5.0
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Devices.Radios==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Media.Control==3.2.1
yarg==0.1.10
zipp==3.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0